tinyice: more fixes

+++ tinyice/tinyice-install.sh
@@ -24,7 +24,7 @@ git-lite \
 go
 
 # Directory Setup
-id -u tinyice 2>&1 || pw user add tinyice -c tinyice -u 8000 -d /nonexistent -s /usr/bin/nologin
+id -u tinyice >/dev/null 2>&1 || pw user add tinyice -c tinyice -u 8000 -d /nonexistent -s /usr/bin/nologin
 mkdir -p /usr/local/etc/tinyice
 mkdir -p /usr/local/etc/rc.d
 mkdir -p /var/run/tinyice
@@ -34,10 +34,13 @@ chown -R tinyice:tinyice /var/log/tinyice
 chown -R tinyice:tinyice /var/run/tinyice
 
 # TinyIce Setup
+if [ "${REINSTALL}" = "true" ]; then
+   service tinyice onestop >/dev/null 2>&1
+   rm -rf /usr/local/tinyice
+fi
 git clone https://github.com/DatanoiseTV/tinyice /usr/local/tinyice
 cd /usr/local/tinyice && make build
 cp -f /usr/local/tinyice/tinyice /usr/local/bin/tinyice
-rm -rf /usr/local/tinyice
 chmod +x /usr/local/bin/tinyice
 fetch -o /usr/local/etc/rc.d/tinyice https://raw.githubusercontent.com/tschettervictor/bsd-apps/main/tinyice/includes/tinyice
 chmod +x /usr/local/etc/rc.d/tinyice
@@ -46,13 +49,19 @@ chmod +x /usr/local/etc/rc.d/tinyice
 sysrc tinyice_enable="YES"
 service tinyice start
 
+# Save Passwords
+SETUP_TOKEN="$(grep 'Setup Token:' /var/log/tinyice/tinyice.log | awk -F': ' '{print $2}')
+echo "${APP_NAME} setup token for first run is ${SETUP_TOKEN}" > /root/${APP_NAME}-Info.txt
+
 echo "---------------"
 echo "Installation complete."
 echo "${APP_NAME} is running on port 8000"
 echo "---------------"
 if [ "${REINSTALL}" != "true" ]; then
-    echo "Note the 'Setup Token' that was just shown."
-    echo "You will need it when first visiting TinyIce."
-    echo "If you missed it, just remove the config file, and restart the service."
+    echo "Setup Information"
+    echo "Setup Token: "${SETUP_TOKEN}""
+    echo "---------------"
+    echo "You will need the Setup Token when running TinyIce for the first time."
+    echo "If you missed it, check /root/${APP_NAME}-Info.txt."
     echo "---------------"
 fi
